Separando o arquivo pecas.py
Separando o arquivo pecas.py em vários arquivos e os colocando na pasta
pecas (também foi criar um __init__.py para facilitar o uso do import)

Usando o typing abc para designar classe abstrata

Aumentando a quantidade de coisas nas classes abstratas para ter
consigurações mais sensatas
