+++ src/pecas/cavalo.py
@@ -1,0 +1,42 @@
+from recursos import Recursos
+from pecas import Peca
+from .util import tabuleiro_false, valida_coordenadas
+
+
+class Cavalo(Peca):
+    def __init__(self, recursos: Recursos, cor: bool):
+        super().__init__(recursos, cor, nome='cavalo')
+
+    def valida_posicao(self, tabuleiro: list[list], pos: tuple[int, int]) -> bool:
+        i, j = pos
+        return tabuleiro[i][j] is None or tabuleiro[i][j].cor != self.cor
+
+    def get_movimentos_simples(self, tabuleiro: list[list], flags: list, pos: tuple[int, int]) -> list[list]:
+        res = tabuleiro_false()
+        i, j = pos
+
+        # Casas acima
+        if valida_coordenadas(i-2, j-1):
+            res[i-2][j-1] = self.valida_posicao(tabuleiro, (i-2, j-1))
+        if valida_coordenadas(i-2, j+1):
+            res[i-2][j+1] = self.valida_posicao(tabuleiro, (i-2, j+1))
+
+        # Casas abaixo
+        if valida_coordenadas(i+2, j-1):
+            res[i+2][j-1] = self.valida_posicao(tabuleiro, (i+2, j-1))
+        if valida_coordenadas(i+2, j+1):
+            res[i+2][j+1] = self.valida_posicao(tabuleiro, (i+2, j+1))
+
+        # Casas a esquerda
+        if valida_coordenadas(i-1, j-2):
+            res[i-1][j-2] = self.valida_posicao(tabuleiro, (i-1, j-2))
+        if valida_coordenadas(i+1, j-2):
+            res[i+1][j-2] = self.valida_posicao(tabuleiro, (i+1, j-2))
+
+        # Casas a direira
+        if valida_coordenadas(i-1, j+2):
+            res[i-1][j+2] = self.valida_posicao(tabuleiro, (i-1, j+2))
+        if valida_coordenadas(i+1, j+2):
+            res[i+1][j+2] = self.valida_posicao(tabuleiro, (i+1, j+2))
+
+        return res
