+++ src/pecas/util.py
@@ -1,0 +1,109 @@
+from copy import copy
+
+from recursos import Recursos
+from pecas import MovimentoEspecial
+
+
+##### Tabuleiro #####
+def tabuleiro_false() -> list[list[bool]]:
+    """
+    :return: list 8x8 com todos os campos sendo False
+    """
+    return [[False] * 8 for _ in range(8)]
+
+
+def tabuleiro_copia(tabuleiro) -> list[list]:
+    copia = [[None] * 8 for _ in range(8)]  # list 8x8 com None
+    for i, linha in enumerate(tabuleiro):
+        for j, peca in enumerate(linha):
+            if peca is not None:
+                copia[i][j] = copy(peca)
+    return copia
+
+##### Xeque #####
+
+
+def testar_xeque(tabuleiro: list[list], flags: list, pos_rei: tuple[int, int]) -> bool:
+    """
+    Testa se o rei está em xeque
+    :param pos_rei: posição do rei
+    """
+    ri, rj = pos_rei
+    rei = tabuleiro[ri][rj]
+    for pi, linha in enumerate(tabuleiro):
+        for pj, peca in enumerate(linha):
+            if peca is not None and peca.cor != rei.cor:
+                movimentos = peca.get_movimentos_simples(
+                    tabuleiro,
+                    flags,
+                    (pi, pj)
+                )
+                if isinstance(movimentos[ri][rj], bool) and movimentos[ri][rj]:
+                    return True
+                elif isinstance(movimentos[ri][rj], MovimentoEspecial) and not movimentos[ri][rj].avanco:
+                    return True
+    return False
+
+
+# TODO pode ser muito otimizado
+def testar_movimento(tabuleiro: list[list], flags: list, recursos: Recursos, pos_rei: tuple[int, int], acao: tuple[tuple[int, int], tuple[int, int]]) -> bool:
+    tab = tabuleiro_copia(tabuleiro)
+    pos, nova_pos = acao
+    i, j = pos
+    m, n = nova_pos
+    movimento = tab[i][j].get_movimentos_simples(tab, flags, pos)[m][n]
+
+    # Movimenta peca
+    if isinstance(movimento, bool) and movimento:
+        tab[m][n] = tabuleiro[i][j]
+        tab[i][j] = None
+        tab[m][n].notifica_movimento()
+
+        if pos_rei == pos:
+            pos_rei = nova_pos
+
+        flags.clear()
+
+    elif isinstance(movimento, MovimentoEspecial):
+        movimento.executar(tab, flags, recursos)
+
+        if movimento.nome == 'roque' and pos_rei == movimento.rei:
+            pos_rei = movimento.nova_rei
+
+        flags.clear()
+        movimento.update_flags(flags)
+    else:
+        return False
+
+    # Testa Xeque
+    return not testar_xeque(tab, flags, pos_rei)
+
+
+##### Movimento #####
+def mover_peca(tabuleiro: list[list], pos: tuple[int, int], nova_pos: tuple[int, int]) -> None:
+    i, j = pos
+    m, n = nova_pos
+    tabuleiro[m][n] = tabuleiro[i][j]
+    tabuleiro[m][n].notifica_movimento()
+    tabuleiro[i][j] = None
+
+
+def valida_coordenadas(a: int, b: int = 0) -> bool:
+    """
+    valida_coordenada(a):    Verifica se a é um valor válido para componente de uma coordenada
+    valida_coordenada(a, b): Verifica se (a, b) é uma coordenada válida
+    """
+    return 0 <= a < 8 and 0 <= b < 8
+
+
+def calcula_direcao(res: list[list], tabuleiro: list[list], pos: tuple[int, int], direcoes: tuple[tuple[int, int], ...], cor: bool) -> None:
+    for (di, dj) in direcoes:
+        i, j = pos
+        i, j = i + di, j + dj
+        while valida_coordenadas(i, j):
+            if tabuleiro[i][j] is None:
+                res[i][j] = True
+            else:
+                res[i][j] = cor != tabuleiro[i][j].cor
+                break  # Se a casa não está vazia, não tem porquê olhar adiante
+            i, j = i + di, j + dj
